Make coin values data driven, fix some coin overflows

+++ src/main/java/gay/nyako/nyakomod/mixin/ServerPlayerEntityMixin.java
@@ -114,7 +114,7 @@ public abstract class ServerPlayerEntityMixin implements ServerPlayerEntityAcces
     public void playerTick(CallbackInfo info) {
         var player = ((ServerPlayerEntity) (Object) this);
         var count = CunkCoinUtils.countInventoryCoins(player.getInventory()) + CunkCoinUtils.countInventoryCoins(player.getEnderChestInventory());
-        player.getScoreboard().forEachScore(NyakoMod.COIN_CRITERIA, player.getEntityName(), score -> score.setScore(count));
+        player.getScoreboard().forEachScore(NyakoMod.COIN_CRITERIA, player.getEntityName(), score -> score.setScore((int) count));
     }
 
     @Redirect(method = "trySleep", at = @At(value = "INVOKE", target = "Lnet/minecraft/world/dimension/DimensionType;natural()Z"))
